# finance-manager-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API serves personal finance endpoints. Accuracy is not garunteed",
        "title": "Go Finance Manager API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/authenticate": {
            "post": {
                "description": "Attempts to use passed credentials to authenticate with the application and generate JWT tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authentication.TokenPairs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
...
        "/refresh": {
            "get": {
                "description": "Attempts to refresh Tokens using a refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authentication.TokenPairs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{userId}/stocks": {
            "get": {
                "description": "Gets a summary of all stock data for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Summary"
                ],
                "summary": "Get Stock Portfolio Summary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserStockPortfolioSummary"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts a new Stock into the Database for a given user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Insert Stock",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The stock to insert",
                        "name": "stock",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserStock"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/summary": {
            "get": {
                "description": "Gets a summary of all financial data for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Summary"
                ],
                "summary": "Get Finance Summary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Summary"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "authentication.TokenPairs": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "jsonutils.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.CreditSummary": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                },
                "utilization": {
                    "type": "number"
                }
            }
        },
...
        "models.ExpenseSummary": {
            "type": "object",
            "properties": {
                "bills": {
                    "type": "number"
                },
                "creditCardBalance": {
                    "type": "number"
                },
                "creditCards": {
                    "type": "number"
                },
                "expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "loanBalance": {
                    "type": "number"
                },
                "loanCost": {
                    "type": "number"
                },
                "overallBalance": {
                    "type": "number"
                },
                "taxes": {
                    "type": "number"
                },
                "totalBalance": {
                    "type": "number"
                },
                "totalCost": {
                    "type": "number"
                }
            }
        },
...
        "models.IncomeSummary": {
            "type": "object",
            "properties": {
                "incomes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummaryItem"
                    }
                },
                "totalIncome": {
                    "type": "number"
                }
            }
        },
...
        "models.PortfolioPosition": {
            "type": "object",
            "properties": {
                "asOf": {
                    "type": "string"
                },
                "close": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "history": {
                    "$ref": "#/definitions/models.PositionHistory"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "quantity": {
                    "type": "number"
                },
                "ticker": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "models.PositionHistory": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "endDt": {
                    "type": "string"
                },
                "startDt": {
                    "type": "string"
                },
                "ticker": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Stock"
                    }
                }
            }
        },
...
        "models.Stock": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "number"
                },
                "createDt": {
                    "type": "string"
                },
                "currentHigh": {
                    "type": "number"
                },
                "currentLow": {
                    "type": "number"
                },
                "currentPrice": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastUpdateDt": {
                    "type": "string"
                },
                "open": {
                    "type": "number"
                },
                "ticker": {
                    "type": "string"
                }
            }
        },
        "models.Summary": {
            "type": "object",
            "properties": {
                "creditSummary": {
                    "$ref": "#/definitions/models.CreditSummary"
                },
                "expenseSummary": {
                    "$ref": "#/definitions/models.ExpenseSummary"
                },
                "incomeSummary": {
                    "$ref": "#/definitions/models.IncomeSummary"
                },
                "netFunds": {
                    "type": "number"
                }
            }
        },
        "models.SummaryItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "balance": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.UserStock": {
            "type": "object",
            "properties": {
                "createDt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastUpdateDt": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "ticker": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "models.UserStockPortfolioSummary": {
            "type": "object",
            "properties": {
                "asOf": {
                    "type": "string"
                },
                "currentClose": {
                    "type": "number"
                },
                "currentHigh": {
                    "type": "number"
                },
                "currentLow": {
                    "type": "number"
                },
                "currentOpen": {
                    "type": "number"
                },
                "currentValue": {
                    "type": "number"
                },
                "positions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PortfolioPosition"
                    }
                }
            }
        }
    }
}